Crate subcoin_service

Subcoin service implementation. Specialized wrapper over substrate service.

Modules

chain_spec

network_request_handler

Helper for answering the subcoin specific requests from a remote peer via the request-response protocol.

transaction_pool

Structs

CoinStorageKey

This struct implements the trait subcoin_primitives::CoinStorageKey.

GenesisBlockBuilder

Subcoin genesis block builder.

NodeComponents

Subcoin node components.

SubcoinConfiguration

Subcoin node configuration.

SubcoinExecutorDispatch

Subcoin executor.

SubcoinFinalizer

This struct is responsible for finalizing blocks with enough confirmations.

TransactionAdapter

Responsible for doing the conversion between Bitcoin transaction and Substrate extrinsic.

Functions

initialize_genesis_block_hash_mapping

Insert the genesis block hash mapping into aux-db.

new_node

Creates a new subcoin node.

new_partial

Creates a partial node, for the chain ops commands.

start_substrate_network

Runs the Substrate networking.

Type Aliases

ChainSpec

This is a specialization of the general Substrate ChainSpec type.

FullBackend

FullClient

Disk backend client type.